+++ src/ts/store.ts
@@ -32,7 +32,7 @@ namespace CdvPurchase {
     /**
      * Current release number of the plugin.
      */
-    export const PLUGIN_VERSION = '13.11.0';
+    export const PLUGIN_VERSION = '13.11.1';
 
     /**
      * Entry class of the plugin.